Fixed trailing comma bug
Depending on the options specified, there could be a trailing comma which is invalid JSON and prevented consul from starting. Lame but easy solution was to make a required configuration be the last one and never have a trailing comma and made sure all other configs do end in a comma.