sudo and set file ownership explicitely
As of ansible 2.1, using an unprivileged user (`become` and
`become_user`) creates an exception since it does not want to create
the temp file for the command with world writeable permissions (see
https://docs.ansible.com/ansible/become.html#becoming-an-unprivileged-us
er for more details). There are many solutions, however just sudo and
set the permissions does the trick.

Fixes savagegus/ansible-consul#146